Move to async TUI

+++ archinstall/scripts/guided.py
@@ -1,5 +1,7 @@
 import os
+import sys
 import time
+from typing import Any
 
 from archinstall.lib.applications.application_handler import ApplicationHandler
 from archinstall.lib.args import ArchConfigHandler
@@ -10,6 +12,7 @@ from archinstall.lib.disk.utils import disk_layouts
 from archinstall.lib.global_menu import GlobalMenu
 from archinstall.lib.installer import Installer, accessibility_tools_in_use, run_custom_user_commands
 from archinstall.lib.interactions.general_conf import PostInstallationAction, select_post_installation
+from archinstall.lib.menu.util import delayed_warning
 from archinstall.lib.mirrors import MirrorListHandler
 from archinstall.lib.models import Bootloader
 from archinstall.lib.models.device import DiskLayoutType, EncryptionType
@@ -19,6 +22,8 @@ from archinstall.lib.output import debug, error, info
 from archinstall.lib.packages.util import check_version_upgrade
 from archinstall.lib.profile.profiles_handler import profile_handler
 from archinstall.lib.translationhandler import tr
+from archinstall.tui.ui.components import tui
+from archinstall.tui.ui.result import Result, ResultType
 
 
 def show_menu(
@@ -42,7 +47,9 @@ def show_menu(
 	if not arch_config_handler.args.advanced:
 		global_menu.set_enabled('parallel_downloads', False)
 
-	global_menu.run()
+	result: Any | None = tui.run(global_menu)
+	if result is None:
+		sys.exit(0)
 
 
 def perform_installation(
@@ -177,18 +184,20 @@ def perform_installation(
 
 		if not arch_config_handler.args.silent:
 			elapsed_time = time.monotonic() - start_time
-			action = select_post_installation(elapsed_time)
+			result: Result[PostInstallationAction] = tui.run(lambda: select_post_installation(elapsed_time))
 
-			match action:
-				case PostInstallationAction.EXIT:
-					pass
-				case PostInstallationAction.REBOOT:
-					os.system('reboot')
-				case PostInstallationAction.CHROOT:
-					try:
-						installation.drop_to_shell()
-					except Exception:
+			if result.type_ == ResultType.Selection:
+				action = result.get_value()
+				match action:
+					case PostInstallationAction.EXIT:
 						pass
+					case PostInstallationAction.REBOOT:
+						_ = os.system('reboot')
+					case PostInstallationAction.CHROOT:
+						try:
+							installation.drop_to_shell()
+						except Exception:
+							pass
 
 
 def main(arch_config_handler: ArchConfigHandler | None = None) -> None:
@@ -212,7 +221,9 @@ def main(arch_config_handler: ArchConfigHandler | None = None) -> None:
 
 	if not arch_config_handler.args.silent:
 		aborted = False
-		if not config.confirm_config():
+		res: bool = tui.run(config.confirm_config)
+
+		if not res:
 			debug('Installation aborted')
 			aborted = True
 
@@ -221,6 +232,11 @@ def main(arch_config_handler: ArchConfigHandler | None = None) -> None:
 
 	if arch_config_handler.config.disk_config:
 		fs_handler = FilesystemHandler(arch_config_handler.config.disk_config)
+
+		rc: Result[bool] = tui.run(lambda: delayed_warning(tr('Starting device modifications in ')))
+		if not rc.get_value():
+			return main()
+
 		fs_handler.perform_filesystem_operations()
 
 	perform_installation(
